Refactor makestyles into separate files, create custom hook for todo list states
