+++ client/src/components/ToDoList/ToDoItem/styles.js
@@ -1,0 +1,13 @@
+import { makeStyles } from "@material-ui/core";
+
+const useStyles = makeStyles({
+    todoCheckBox: {
+        color: "#30A0F5"
+    },
+    subItem: {
+        paddingTop: 0,
+        paddingBottom: 0
+    },
+});
+
+export default useStyles;
