+++ client/src/components/ToDoList/TodaysToDo/index.js
@@ -1,17 +1,9 @@
 import React from "react";
 import PropTypes from "prop-types";
-import { CardActions, Fab, List, makeStyles } from "@material-ui/core";
+import { CardActions, Fab, List } from "@material-ui/core";
 import AddIcon from '@material-ui/icons/Add';
 import ToDoItem from "../ToDoItem";
-
-const useStyles = makeStyles({
-    fab: {
-        backgroundColor: "#30A0F5"
-    },
-    cardActions: {
-        justifyContent: "flex-end"
-    }
-});
+import useStyles from "./styles";
 
 function TodaysToDo({ todaysDate, todoList, switchToAdd, toggleCheck, deleteItem, editItem }) {
     const classes = useStyles();
