+++ client/src/components/ToDoList/UpcomingToDo/index.js
@@ -1,18 +1,10 @@
 import React from "react";
 import PropTypes from "prop-types";
-import { CardActions, Fab, List, ListSubheader, makeStyles } from "@material-ui/core";
+import { CardActions, Fab, List, ListSubheader } from "@material-ui/core";
 import AddIcon from '@material-ui/icons/Add';
 import ToDoItem from "../ToDoItem";
 import moment from "moment";
-
-const useStyles = makeStyles({
-    fab: {
-        backgroundColor: "#30A0F5"
-    },
-    cardActions: {
-        justifyContent: "flex-end"
-    }
-});
+import useStyles from "./styles";
 
 function UpcomingToDo({ upcomingToDoList, switchToAdd, toggleCheck, deleteItem, editItem }) {
     const classes = useStyles();
