+++ client/src/components/ToDoList/UpcomingToDo/styles.js
@@ -1,0 +1,12 @@
+import { makeStyles } from "@material-ui/core";
+
+const useStyles = makeStyles({
+    fab: {
+        backgroundColor: "#30A0F5"
+    },
+    cardActions: {
+        justifyContent: "flex-end"
+    }
+});
+
+export default useStyles;
